+++ lib/light/arkit_light_type.dart
@@ -1,0 +1,9 @@
+/// Describes the various light types available.
+enum ARKitLightType {
+  ambient,
+  omni,
+  directional,
+  spot,
+  ies,
+  probe,
+}